add more tests + core is a single file again
