+++ tests/__main__.py
@@ -1,7 +1,7 @@
 # Copyright © 2023- Frello Technology Private Limited
 
 from tests.getkv import TestGetValueByKeys
-from tests.chains import TestChainSerDeser
+from tests.base import TestSerDeser, TestNode
 import unittest
 
 if __name__ == "__main__":
